+++ R/flux_match.R
@@ -39,6 +39,7 @@
 #' @importFrom tidyr fill drop_na
 #' @importFrom lubridate is.POSIXct
 #' @importFrom lifecycle deprecate_stop deprecated deprecate_warn is_present
+#' @importFrom rlang as_label enquo
 #' @examples
 #' data(co2_df_short, record_short)
 #' flux_match(co2_df_short, record_short, datetime, start,
@@ -90,8 +91,8 @@ flux_match <- function(raw_conc,
     )
   }
 
-  name_raw_conc <- deparse(substitute(raw_conc))
-  name_field_record <- deparse(substitute(field_record))
+  name_raw_conc <- as_label(enquo(raw_conc))
+  name_field_record <- as_label(enquo(field_record))
 
   args_ok <- flux_fun_check(
     list(time_diff = time_diff),
@@ -118,10 +119,15 @@ flux_match <- function(raw_conc,
   if (any(!c(args_ok, raw_conc_ok, field_record_ok)))
     stop("Please correct the arguments", call. = FALSE)
 
+  # test if cols in field_record and raw_conc are different
+  colnames_raw_conc <- names(raw_conc)
+  colnames_field_record <- names(field_record)
 
+  colnames_test <- colnames_raw_conc %in% colnames_field_record
 
-
-
+  if (any(colnames_test)) {
+    stop("raw_conc and field_record must have different column names")
+  }
 
 
   field_record <- field_record |>