+++ R/flux_plot.R
@@ -27,16 +27,18 @@
 #' @param y_text_position position of the text box
 #' @param print_plot logical, if TRUE it prints the plot as a ggplot object
 #' but will take time depending on the size of the dataset
-#' @param output `pdfpages`, the plots are saved as A4 landscape pdf pages;
-#' `ggsave`, the plots can be saved with the ggsave function;
-#' `print_only` (default) prints the plot without creating a file
-#' (independently from `print_plot` being TRUE or FALSE)
+#' @param output `"pdfpages"`, the plots are saved as A4 landscape pdf pages;
+#' `"ggsave"`, the plots can be saved with the ggsave function;
+#' `"print_only"` (default) prints the plot without creating a file
+#' (independently from `print_plot` being TRUE or FALSE);
+#' `"longpdf"`, the plots are saved as a pdf file as long as needed (faster than
+#' `"pdfpages"`)
 #' @param ggsave_args list of arguments for \link[ggplot2:ggsave]{ggsave}
 #' (in case `output = "ggsave"`)
 #' @param f_facetid character vector of columns to use as facet IDs. Note that
 #' they will be united, and that has to result in a unique facet ID for each
 #' measurement. Default is `f_fluxid`
-#' @param longpdf_args arguments for plotly in the form
+#' @param longpdf_args arguments for longpdf in the form
 #' `list(ncol, width (in cm), ratio)`
 #' @return plots of fluxes, with raw concentration data points, fit, slope,
 #' and color code indicating quality flags and cuts. The plots are organized
@@ -121,7 +123,7 @@ flux_plot <- function(slopes_df,
 
 
   if (f_plotname == "") {
-    f_plotname <- deparse(substitute(slopes_df))
+    f_plotname <- as_label(enquo(slopes_df))
   }
 
   if (output %in% c("pdfpages", "ggsave", "longpdf")) {
@@ -152,7 +154,7 @@ flux_plot <- function(slopes_df,
     )
 
   if (
-    max(slopes_df[[deparse(substitute(f_conc))]], na.rm = TRUE) > f_ylim_upper
+    max(slopes_df[[as_label(enquo(f_conc))]], na.rm = TRUE) > f_ylim_upper
   ) {
     message("Some concentration data points will not be displayed
     because f_ylim_upper is too low.")
@@ -164,7 +166,7 @@ flux_plot <- function(slopes_df,
   }
 
   if (
-    min(slopes_df[[deparse(substitute(f_conc))]], na.rm = TRUE) < f_ylim_lower
+    min(slopes_df[[as_label(enquo(f_conc))]], na.rm = TRUE) < f_ylim_lower
   ) {
     message("Some concentration data points will not be displayed
     because f_ylim_lower is too high.")
@@ -261,7 +263,7 @@ flux_plot <- function(slopes_df,
   f_plot <- f_plot +
     geom_line(
       aes(y = .data$f_fit, linetype = .data$linetype),
-      linewidth = 0.3,
+      linewidth = 0.5,
       na.rm = TRUE,
       show.legend = TRUE
     ) +
@@ -292,7 +294,7 @@ flux_plot <- function(slopes_df,
       colour = "Quality flags",
       linetype = "Fits"
     ) +
-    guides(color = guide_legend(override.aes = list(linetype = 0)))
+    guides(color = guide_legend(override.aes = list(linetype = 0, size = 3)))
 
 
   if (output == "pdfpages") {